backend/requirements.txt
apturl==0.5.2
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
bcrypt==3.2.2
blinker==1.4
Brlapi==0.8.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
colorclass==2.2.2
command-not-found==0.3
cryptography==37.0.2
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.0.4
django-cors-headers==3.12.0
django-countries==7.3.2
django-email-verification==0.2.2
djangorestframework==3.13.1
docopt==0.6.2
duplicity==0.8.21
executing==0.8.3
fasteners==0.17.3
future==0.18.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
ipython==8.3.0
jedi==0.18.1
jeepney==0.8.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
monotonic==1.6
more-itertools==8.13.0
mysqlclient==2.1.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.10.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pip-upgrader==1.4.15
prompt-toolkit==3.0.29
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.12.0
PyGObject==3.42.0
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
reportlab==3.6.9
requests==2.27.1
SecretStorage==3.3.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
systemd-python==234
terminaltables==3.1.10
traitlets==5.2.0
typing_extensions==4.2.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.9
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
xdg==5.1.1
xkit==0.0.0
zipp==3.8.0
